セーブデータクラスのサンプルを追加
