+++ Assets/ExampleUsing/EncryptedSave.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 185a58c58ad988648956829f63f5f7fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
