+++ Assets/ExampleUsing/EncryptedSave/EncryptedSave.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d681120243d1f4f4d960c4b56e747c69
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
