+++ Assets/ExampleUsing/EncryptedSave/EncryptedSaveExample.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1bf09e08cfa47f4fa8c9562af1892a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
