+++ Assets/TakiExtension/EncryptedSave.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d98794e8c916cdf49b2484aaf328c3a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
