+++ Assets/TakiExtension/EncryptedSave/Save.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 24b2c74d75f8b7b4480299643ffa855f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
