+++ Assets/TakiExtension/EncryptedSave/SaveDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b511cd4bb7131f41a2d5facc3b0bddc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
