--- Assets/TakiExtension/Save.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d98794e8c916cdf49b2484aaf328c3a4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
